requirements.txt
fastapi==0.115.12
faiss-cpu==1.11.0
uvicorn==0.34.2
google-genai==1.14.0
numpy==2.2.4
python-multipart==0.0.20